--- a/docs/cmdline-opts/trace-config.d
+++ b/docs/cmdline-opts/trace-config.d
@@ -16,6 +16,6 @@ where detailed output will be made available from. Names are case-insensitive.
 Specify 'all' to enable all trace components.
 
 In addition to trace component names, specify "ids" and "time" to
-avoid extra --trace-ids or --trace-time commandline parameters.
+avoid extra --trace-ids or --trace-time parameters.
 
 See the *curl_global_trace(3)* man page for more details.