--- a/docs/cmdline-opts/trace-config.d
+++ b/docs/cmdline-opts/trace-config.d
@@ -8,9 +8,14 @@ Category: verbose
 Example: --trace-config ids,http/2 $URL
 Added: 8.3.0
 See-also: verbose trace
-Multi: single
+Multi: append
 Scope: global
 ---
 Set configuration for trace output. A comma-separated list of components
-where detailed output will be made available. Specify 'all' to enable all
-details.
+where detailed output will be made available from. Names are case-insensitive.
+Specify 'all' to enable all trace components.
+
+In addition to trace component names, specify "ids" and "time" to
+avoid extra --trace-ids or --trace-time commandline parameters.
+
+See the *curl_global_trace(3)* man page for more details.