--- a/docs/libcurl/curl_global_log_config.3
+++ b/docs/libcurl/curl_global_log_config.3
@@ -31,7 +31,7 @@ curl_global_log_config - Global libcurl logging configuration
 CURLcode curl_global_log_config(const char *config);
 .fi
 .SH DESCRIPTION
-This function configures the logging behaviour, allowing to make some
+This function configures the logging behavior, allowing to make some
 parts of curl more verbose or silent than others.
 
 This function may be called during the initialization phase of a program. It