--- a/docs/libcurl/curl_global_trace.3
+++ b/docs/libcurl/curl_global_trace.3
@@ -56,19 +56,38 @@ If you are initializing libcurl from a Windows DLL you should not initialize
 it from \fIDllMain\fP or a static initializer because Windows holds the loader
 lock during that time and it could cause a deadlock.
 
-The \fIconfig\fP string is a list of comma-separated component names. Unknown
-names are ignored. The special name "all" applies to all components. Names
-may be prefixed with '+' or '-' to enable or disable detailed logging for
-a component.
+The \fIconfig\fP string is a list of comma-separated component names. Names
+are case-insensitive and unknown names are ignored. The special name "all"
+applies to all components. Names may be prefixed with '+' or '-' to enable
+or disable detailed logging for a component.
 
 The list of component names is not part of curl's public API. Names may
 be added or disappear in future versions of libcurl. Since unknown names
 are silently ignored, outdated log configurations will not error when
-upgrading libcurl.
+upgrading libcurl. Given that, some names can be expected to be fairly
+stable and are listed below for easy reference.
 
 Note that log configuration applies only to transfers where debug logging
 is enabled. See \fICURLOPT_VERBOSE(3)\fP or \fICURLOPT_DEBUGFUNCTION(3)\fP
 on how to control that.
+
+.SH TRACE COMPONENTS
+.IP tcp
+Tracing of TCP socket handling: connect, reads, writes.
+.IP ssl
+Tracing of SSL/TLS operations, whichever SSL backend is used in your build.
+.IP http/2
+Details about HTTP/2 handling: frames, events, I/O, etc.
+.IP http/3
+Details about HTTP/3 handling: connect, frames, events, I/O etc.
+.IP http-proxy
+Involved when transfers are tunneled through a HTTP proxy. "h1-proxy" or
+"h2-proxy" are also involved, depending on the HTTP version negotiated with
+the proxy.
+
+In order to find out all components involved in
+a transfer, run it with "all" configured. You will then see all names
+involved in your libcurl version in the trace.
 
 .SH EXAMPLE
 .nf
@@ -77,8 +96,18 @@ on how to control that.
 
  /* log all details, except SSL handling */
  curl_global_trace("all,-ssl");
-
 .fi
+
+Below is a trace sample where "http/2" was configured. The trace output
+of an enabled component appears at the beginning in brackets.
+.nf
+* [HTTP/2] [h2sid=1] cf_send(len=96) submit https://example.com/
+...
+* [HTTP/2] [h2sid=1] FRAME[HEADERS]
+* [HTTP/2] [h2sid=1] 249 header bytes
+...
+.fi
+
 .SH AVAILABILITY
 Added in 8.3
 .SH RETURN VALUE