--- a/include/curl/curl.h
+++ b/include/curl/curl.h
@@ -2732,18 +2732,18 @@ CURL_EXTERN CURLcode curl_global_init_mem(long flags,
 CURL_EXTERN void curl_global_cleanup(void);
 
 /*
- * NAME curl_global_log_config()
+ * NAME curl_global_trace()
  *
  * DESCRIPTION
  *
- * curl_global_log_config() can be invoked at application start to
- * configure which components in curl should participate in logging.
+ * curl_global_trace() can be invoked at application start to
+ * configure which components in curl should participate in tracing.
 
  * This function is thread-safe if CURL_VERSION_THREADSAFE is set in the
  * curl_version_info_data.features flag (fetch by curl_version_info()).
 
  */
-CURL_EXTERN CURLcode curl_global_log_config(const char *config);
+CURL_EXTERN CURLcode curl_global_trace(const char *config);
 
 /* linked-list structure for the CURLOPT_QUOTE option (and other) */
 struct curl_slist {