… lib/cf-h2-proxy.c
… lib/cf-https-connect.c
… lib/cf-socket.c
… lib/connect.c
… lib/curl_log.h
… lib/curl_trc.c
… lib/curl_trc.h
… lib/http2.c
… lib/http_proxy.c
… lib/url.c
… lib/urldata.h
… lib/vquic/curl_msh3.c
… lib/vquic/curl_ngtcp2.c
… lib/vquic/curl_quiche.c
… lib/vquic/vquic.c
… lib/vssh/libssh2.c
… lib/vtls/mbedtls.c
… lib/vtls/openssl.c
… lib/vtls/rustls.c
… lib/vtls/sectransp.c
… lib/vtls/vtls.c
… lib/vtls/wolfssl.c
… libcurl.def
… src/tool_getparam.c
… src/tool_listhelp.c
… tests/data/test1135
… tests/http/test_15_tracing.py
… tests/http/testenv/curl.py
… tests/unit/unit2600.c
… tests/unit/unit2601.c