Making detailed logging availabel in libcurl and curl:

- curl_global_log_config() for setting log configration in libcurl
- `- --trace-config` command line option for curl
- man pages
- convert `DEBUGF(LOG_CF(..))` code to `CURL_LOG_CF()` that is available in non-debug builds
- protect `infof()` execution by verbosity check, saving some  cycles when not verbose
- test adjustments for new function

Examples:

Add HTTP/2 details to verbose output:

            curl -v --trace-config http/2 URL

Add HTTP/2, SSL and id details to verbose output:

            curl -v --trace-config http/2,ssl,ids URL

Enable all details in verbose output, including times and ids and all components in libcurl:

            curl -v --trace-config all URL

Enable all details, but not about SSL or TCP:

            curl -v --trace-config all,-ssl,-tcp URL