@bagder 2023-08-01
> We'll need to update `libcurl.def` with the new API:

Can't we generate that list? Or at worst make a CI job that checks that the list is up-to-date?