@vszakats 2023-08-01
@bagder PR ready for checking it: #11570. I can't imagine a way to dynamically generate this from both CMake and `Makefile.mk` (and perhaps others that may be using this in the future). Converting it a raw list (without the `EXPORTS` header) and re-using it for more than `.def` purposes might work, so that a single list needs to be maintained, but it's non-trivial.