@icing 2023-08-02
> ```diff
> curl_global_log_config
> ```

Added this in the recent push. Thanks for noticing