@icing 2023-08-02
Some local measurement on my macOS dev machine. I see no real difference in performance between non-debug builds on master and this PR. Meaning the guarded log statements have no impact.

```
curl 8.3.0-DEV (x86_64-apple-darwin22.5.0) libcurl/8.3.0-DEV OpenSSL/3.0.9 zlib/1.2.11 brotli/1.0.9 zstd/1.5.5 c-ares/1.19.1 libidn2/2.3.4 libssh2/1.11.0 nghttp2/1.55.0 ngtcp2/0.17.0 nghttp3/0.13.0 librtmp/2.3

Downloads from httpd:
> python3 tests/http/scorecard.py -d --httpd h2
              Size       Single       Serial     Parallel
master:      100MB     703 MB/s     972 MB/s     934 MB/s
this PR:     100MB     710 MB/s     986 MB/s     959 MB/s

Small requests:
> python3 tests/http/scorecard.py -r --httpd h2

Requests, max in parallel
         Size        1            6           25           50          100
master:  10KB     1665 r/s     3428 r/s     3537 r/s     3522 r/s     3475 r/s
this PR: 10KB     1676 r/s     3439 r/s     3536 r/s     3515 r/s     3227 r/s

```