@icing 2023-08-02
> Thought about how this can be tested, at least rudimentary?

Added test cases in pytest.